==> 03_Appendix.Rmd <==
---
title: "03 Appendix Tables"
output:
  html_document:
    df_print: paged
---

## Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(pacman)
# needed for first time installing map packages
#install.packages("rgeos")
pacman::p_load("tidyverse", 
               "ggpubr", 
               "ragg", 
               "kableExtra", 
               "ggmap", 
               "countrycode")

knitr::opts_chunk$set(message = F, warning = F)

```


## Load Data

```{r load dfs}
#df <- readRDS(here::here("data","df.RDS"))
df_na <- readRDS(here::here("data", "df_na.RDS"))
df_macro <- readRDS(here::here("data", "df_macro.RDS"))

```


## Descriptives

### Table X. Cases by Country-Wave

```{r tbl1, include = T, echo = T}
# Cross-tab years and countries
options(knitr.kable.NA = '')
tab1 <- df_na %>%
  dplyr::group_by(country, wave) %>%
  dplyr::summarize(n=n()) %>%
  spread(wave, n) 

tab1 %>%
  kable(., 
        #col.names = c("Country", "1987", "1992", "1999", "2009", "2019"), 
        caption = "Country Sample Size by Wave, ISSP 'Social Inequality (SI)", format.args = list(big.mark = ",")) %>%
kable_styling()

write.csv2(tab1, here::here("results", "country_wave_N.csv"))
```
### Table 2. Descriptives w/ Question Wording

```{r wording}
liberal_redist_label <- "Government should take more responsibility to ensure that everyone is provided for [versus] People should take more responsibility to provide for themselves"
liberal_label <- "Private ownership of business and industry should be increased [versus] government ownership..."
confidence_gov_label <- "I am going to name a number of organizations. For each one, could you tell me how much confidence you have in them: is it a great deal of confidence, quite a lot of confidence, not very much confidence or none at all? [the government]"
democratic_label <- "And how democratically is this country being governed today? Again using a scale from 1 to 10, where 1 means that it is “not at all democratic” and 10 means that it is “completely democratic,” what position would you choose?"
```

### Macro-Correlations


```{r var_dvs}
# create a variable that captures the percentage of respondents that think income gap is too large but do not agree gov should do something about it


df_macro_corr <- df_macro %>%
  dplyr::select(-c(iso3c, wave)) %>%
  cor(., use = "pairwise.complete.obs") %>%
  as.data.frame() %>%
  round(.,2)
```

#### Puzzling correlation check

The positive correlation of liberal_redist and gov_redist is puzzling. Why would a country have more support for people (versus the govt) taking more responsibility to provide for their own needs and agree that the government should reduce income differences?

To investigate we plot the countries, and the result makes clear that this variable cannot be coded correctly. There is no way that the USA is the most supportive of all countries of the government (versus people themselves) taking responsibility for providing for their needs. Yet the WVS codebook suggets that lower values are more support of government.

This is now fixed in the source load for the WVS so re-creating the plot will not reveal the error.

NOTE: We do not use 'liveral values' for this paper anymore, but leavea the variable in for future research.

```{r invest_corr}

# df_macro %>%
#   ggplot(aes(y = gov_redist, x = liberal_redist_i)) +
#   geom_label(aes(label = iso3c))

```

### Table 1. Descriptives

```{r t1desc}
tbl1 <- as.data.frame(matrix(nrow = 22, ncol = 7))

colnames(tbl1) <- c("Variable", "Mean", "SD", "Min", "Max", "Obs", "Source")
tbl1[,1] <- c("Country-Wave-Level Variables", "YGovRedist", "YIncDiff", "Liberal Values I, raw", "Liberal Values I, interp", "Liberal Values II, raw", "Liberal Values II, interp", "Lack of Trust, raw", "Lack of Trust, interp", "Corruption Perceptions, raw", "Corruption Perceptions, interp", "GDP per capita", "Social Spending", "Percent Foreign-Born", "Gini, Disposable", "Waves per Country", "Individual-Level-Variables", "YGovRedist", "YIncDiff", "Age in Years", "Female", "Education in Years")

# get summary stats
tbl1c1 <- df_macro %>%
  select(gov_redist, incdiff_large, lib_redistZ, lib_redistZ_i, libZ, libZ_i, noconf_govZ, noconf_govZ_i, cpiZ, cpiZ_i, gdp_pc_10k, socx, pct_fb_i, gini_disp) %>%
  summarise_all(.funs = c("mean", "sd", "min", "max"), na.rm = T) %>%
  round(.,3)

tbl1[2:15,2] <- t(tbl1c1[1,1:14])
tbl1[2:15,3] <- t(tbl1c1[1,15:28])
tbl1[2:15,4] <- t(tbl1c1[1,29:42])
tbl1[2:15,5] <- t(tbl1c1[1,43:56])

# get N
tbl1[2:15,6] <- c(sum(!is.na(df_macro$gov_redist)),
                  sum(!is.na(df_macro$incdiff_large)),
                  sum(!is.na(df_macro$lib_redistZ)),
                  sum(!is.na(df_macro$lib_redistZ_i)),
                  sum(!is.na(df_macro$libZ)),
                  sum(!is.na(df_macro$libZ_i)),
                  sum(!is.na(df_macro$noconf_govZ)),
                  sum(!is.na(df_macro$noconf_govZ_i)),
                  sum(!is.na(df_macro$cpiZ)),
                  sum(!is.na(df_macro$cpiZ_i)),
                  sum(!is.na(df_macro$gdp_pc_10k)),
                  sum(!is.na(df_macro$socx)),
                  sum(!is.na(df_macro$pct_fb_i)),
                  sum(!is.na(df_macro$gini_disp)))

# countries per wave
tbl1[16,2:5] <- df_macro %>%
  group_by(iso3c) %>%
  mutate(count = n()) %>%
  summarise_all(.funs = "mean") %>%
  select(count) %>%
  summarise_all(.funs = c("mean", "sd", "min", "max"))

# individual level
tbl1c1 <- df_na %>%
  select(gov_redist, incdiff_large, ageC, female, educyrs) %>%
  summarise_all(.funs = c("mean", "sd", "min", "max"), na.rm = T) %>%
  round(.,3)

tbl1[18:22,2] <- t(tbl1c1[1,1:5])
tbl1[18:22,3] <- t(tbl1c1[1,6:10])
tbl1[18:22,4] <- t(tbl1c1[1,11:15])
tbl1[18:22,5] <- t(tbl1c1[1,16:20])

# individual n's
tbl1[18:22,6] <- c(sum(!is.na(df_na$gov_redist)),
                   sum(!is.na(df_na$incdiff_large)),
                   sum(!is.na(df_na$ageC)),
                   sum(!is.na(df_na$female)),
                   sum(!is.na(df_na$educyrs))
                   )

tbl1[2:22,7] <- c("ISSP", "ISSP", "IVS", "IVS", "IVS", "IVS", "IVS", "IVS", "CPI", "CPI", "V-Dem", "OECD", "World Bank", "Solt", "", "", "ISSP", "ISSP", "ISSP", "ISSP", "ISSP")

write.csv(tbl1, file = here::here("results", "tbl1.csv"), row.names = F)
```


## Figure 1

### Prep Data

```{r f1prep}
df_tbl <- df_macro %>%
  select(iso3c, gov_redist_agree, gap_no_gov, corr, incdiff_large_agree) %>%
  group_by(iso3c) %>%
  summarise_all(mean, na.rm = T) %>%
  ungroup()

df_tbl <- df_tbl %>%
  mutate(country = countrycode(iso3c, "iso3c", "country.name")) %>%
  select(country, iso3c, incdiff_large_agree, gov_redist_agree, gap_no_gov, corr) %>%
  mutate(incdiff_large_agree = round(incdiff_large_agree,1),
     	gov_redist_agree = round(gov_redist_agree,1),
     	gap_no_gov = round(gap_no_gov,1),
     	corr = round(corr,2))

df_tbl <- as.data.frame(df_tbl[order(df_tbl$incdiff_large_agree, decreasing = T),])

# position adjustments for final plot
tops <- c("PHL","LVA","HRV","CYP")
lefts <- c("HUN","ZAF","AUT")
bots <- c(tops,lefts)
topss <- df_tbl %>%
  subset(iso3c %in% tops)
leftss <- df_tbl %>%
  subset(iso3c %in% lefts)
botss <- df_tbl %>%
  subset(!(iso3c %in% bots))
```

### Plot F1

```{r tbl_statt_map}
mid = 0.48

agg_png(filename = here::here("results","Fig1.png"), width = 1200, height = 1100, res = 144)

#new plot
ggplot() + 
  geom_point(data = df_tbl, 
             aes(x = incdiff_large_agree , y = gap_no_gov, color = corr),
             alpha = 0.75, 
             size=4) + 
  #bottom labels
  geom_text(data = botss, 
            label = botss$iso3c, 
            aes(x = incdiff_large_agree , y = gap_no_gov, color = corr),
            position = "dodge",
            vjust=2 
            )+
  #top labels
  geom_text(data = topss, 
            label = topss$iso3c, 
            aes(x = incdiff_large_agree , y = gap_no_gov, color = corr),
            position = "dodge",
            vjust=-1.2 
            )+
  #left labels
  geom_text(data = leftss, 
            label = leftss$iso3c, 
            aes(x = incdiff_large_agree , y = gap_no_gov, color = corr),
            position = "dodge",
            hjust=1.35
            )+
  scale_color_gradient2(midpoint=mid, low="#FDE725FF", mid="#29AF7FFF", high="#440154FF", space="Lab") +
  scale_y_continuous(labels = scales::percent_format(scale = 1),
                     breaks = c(10,20,30,40,50,60,70,80,90,100),
                     limits=c(0,40)) +
  scale_x_continuous(labels = scales::percent_format(scale = 1),
                      breaks = c(40,50,60,70,80,90,100),  
                      limits=c(70,100)) + 
  labs(x = " \nPercentage who agree that income gap is too large\nnot counting those who are neutral",
       y = "Percentage who do not support government redistribution\n out of those who agree income gap is too large\n ",
       color = "Within\nCountry\nCorrelation") +
  theme_classic() + 
  theme(legend.position = "right",
        axis.text=element_text(size=12),
        axis.title=element_text(size=14))

dev.off()

knitr::include_graphics(here::here("results", "Fig1.png"))
```


## Figure 2. Appendix

### Prep Wave Plot

This is not ideal but it is necessary to get ggarrange to plot all 5 waves in the legend because no single country has all five waves

### Pooled Part

```{r corall, message = F, warning = F}
#create plot with regression line and regression equation and r^2
mmmPOOLED <- ggplot() + geom_smooth(data=df_na, aes(x=incdiff_large, y=gov_redist, color = as.character(wave)), size=1, se=F, method="lm", linetype=0) +
  geom_smooth(data=df_na, aes(x=incdiff_large, y=gov_redist, color = "Pooled"),se=T, size=1, method="lm") +
  scale_color_manual( breaks = c("1987", "1992", "1999", "2009", "2019","Pooled"),
                        values = c("#B8DE29FF", "#29AF7FFF", "#2D708EFF", "#453781FF", "#440154FF","#c83349ff")) +
    
        #geom_point() +
        xlim(1,5) + # make both axes the same for easier visualization
        ylim(1,5) +
        #stat_regline_equation(label.x=2.5, label.y=1.6, hjust = 0) +
        #stat_cor(aes(label=..rr.label..), label.x=2.5, label.y=1, hjust = 0) +
        labs(x = "Gov. should reduce income differences", y = "Inequality btw. rich and poor is too large", color = "Wave") +
        annotate(geom = "text", x = 2.5, y = 1.3, label = paste("R = ", round(cor(df_na$gov_redist, df_na$incdiff_large, use = "pairwise.complete.obs"),2)),
                      hjust = 0,
                      ) +
        annotate(geom = "text", x = 1.2, y = 4.5, 
                 label = paste0("Pooled data for ",length(unique(df_na$iso3c[!is.na(df_na$incdiff_large)])),"\ncountries"),
                 hjust = 0) + #add number of countries to plot
        theme_classic2() +
        theme(
          axis.title = element_text(size = 8)
        ) 

mmmPOOLED +
  theme( axis.title = element_text(size = 12), 
    legend.position="none")


```

As we are aware, this regression line (i.e., correlation) is different in different countries. For example the correlation was lower in the US than in Sweden. That means that the $R^2$ is also different because it is really just the correlation (i.e., the regression coefficient $B_{1}$) squared.

But to be clear, $R^2$ is the most common statistic used to identify what Snijders and Bosker refer to as "explained variance" in the first sentence of their article.

Lets look at the association in all 34 countries now.

### Prep Data, Country Part

```{r reg_eacha}
# get all intercepts and slopes
fitted_models = df_na %>% # have to use equal sign here because it is a function
  subset(!is.na(gov_redist) & !is.na(incdiff_large)) %>% # remove missing cases
  group_by(iso3c) %>% # group by country
  do(model = lm(gov_redist ~ incdiff_large, data = .)) #run regression for each country

# plot all at once

# now loop to create a plot for each intercept and coefficient from each country's regression

for (j in 1:length(unique(fitted_models$iso3c))){
  dfx <- df_na %>% # subset data to only that country and remove missing
    subset(iso3c == fitted_models$iso3c[j]) %>% # this is just a shortcut to cycle through each country that has relevant variables
    subset(!is.na(gov_redist) & !is.na(incdiff_large))
  
  plot1 <- ggplot(data = dfx, aes(x = incdiff_large, y = gov_redist, 
                                  #group = wave, 
                                  color=as.character(wave))) +
           geom_smooth(method = 'lm', se=T) +
           xlim(1,5) +
           ylim(1,5) +
    scale_color_manual( breaks = c("1987", "1992", "1999", "2009", "2019"),
                        values = c("#B8DE29FF", "#29AF7FFF", "#2D708EFF", "#453781FF", "#440154FF")) +
           #annotate(geom = "text", x = 2.5, y = 1.6, label = paste0("y = ",
            #                                                  round(fitted_models$model[[j]]$coefficients[1],1),
             #                                                 " + ",
              #                                                round(fitted_models$model[[j]]$coefficients[2],2),"x"), hjust = 0,) +
           annotate(geom = "text", x = 2.5, y = 1.3, label = paste("Average\nR^2 = ", round(cor(dfx$gov_redist, dfx$incdiff_large)^2,2)),
                      hjust = 0,
                      ) +
           #stat_cor(label.x = 2.5, label.y = 1.3, hjust = 0, p.accuracy = 0.001) +
           labs(title = paste0(fitted_models$iso3c[j]), color='Wave') +
           theme_classic2() +
        theme(axis.title = element_blank())
             #theme(axis.title = element_blank(), legend.key.size = unit(2, 'mm'), legend.direction = "horizontal",legend.position="top", legend.justification="right", legend.margin=margin(0,0,0,0))
    
 
assign(paste0("mmm",fitted_models$iso3c[j]), plot1)
  
}
         

```

### Plot F2

```{r reg_each_plota, message = F, warning = F}
# now gather all and plot with ggarrange

agg_png(filename = here::here("results","Fig2.png"), res = 72, width = 1200, height = 1200)
ggarrange(mmmPOOLED, mmmAUS, mmmAUT, mmmBEL, mmmBGR, mmmCAN, mmmCHE, mmmCHL, 
mmmCYP, mmmCZE, mmmDEU, mmmDNK, mmmESP, mmmFIN, mmmFRA, 
mmmGBR, mmmHRV, mmmHUN, mmmIRL, mmmISR, mmmITA, mmmJPN, 
mmmLVA, mmmNLD, mmmNOR, mmmNZL, mmmPHL, mmmPOL, mmmPRT, 
mmmRUS, mmmSVK, mmmSVN, mmmSWE, mmmTHA, mmmUSA, mmmZAF, common.legend = TRUE) 
  
dev.off()


knitr::include_graphics(here::here("results","Fig2.png"))
```

### New Figure Corrupt_micro

```{r reg_each}
# get all intercepts and slopes
fitted_models = df_na %>% # have to use equal sign here because it is a function
  subset(!is.na(gov_redist) & !is.na(corrupt_top_r)) %>% # remove missing cases
  group_by(iso3c) %>% # group by country
  do(model = lm(gov_redist ~ corrupt_top_r, data = .)) #run regression for each country

# plot all at once

# now loop to create a plot for each intercept and coefficient from each country's regression

for (j in 1:length(unique(fitted_models$iso3c))){
  dfx <- df_na %>% # subset data to only that country and remove missing
    subset(iso3c == fitted_models$iso3c[j]) %>% # this is just a shortcut to cycle through each country that has relevant variables
    subset(!is.na(gov_redist) & !is.na(corrupt_top_r))
  
  plot1 <- ggplot(data = dfx, aes(x = corrupt_top_r, y = gov_redist, 
                                  #group = wave, 
                                  color=as.character(wave))) +
           geom_smooth(method = 'lm', se=T) +
           xlim(1,5) +
           ylim(1,5) +
    scale_color_manual( breaks = c("1987", "1992", "1999", "2009", "2019"),
                        values = c("#B8DE29FF", "#29AF7FFF", "#2D708EFF", "#453781FF", "#440154FF")) +
           #annotate(geom = "text", x = 2.5, y = 1.6, label = paste0("y = ",
            #                                                  round(fitted_models$model[[j]]$coefficients[1],1),
             #                                                 " + ",
              #                                                round(fitted_models$model[[j]]$coefficients[2],2),"x"), hjust = 0,) +
           annotate(geom = "text", x = 2.5, y = 1.3, label = paste("Average\nR^2 = ", round(cor(dfx$gov_redist, dfx$incdiff_large)^2,2)),
                      hjust = 0,
                      ) +
           #stat_cor(label.x = 2.5, label.y = 1.3, hjust = 0, p.accuracy = 0.001) +
           labs(title = paste0(fitted_models$iso3c[j]), color='Wave') +
           theme_classic2() +
        theme(axis.title = element_blank())
             #theme(axis.title = element_blank(), legend.key.size = unit(2, 'mm'), legend.direction = "horizontal",legend.position="top", legend.justification="right", legend.margin=margin(0,0,0,0))
    
 
assign(paste0("rrr",fitted_models$iso3c[j]), plot1)
  
}
         

```

### Plot New Corr of Corrupt_micro

```{r reg_each_plot, message = F, warning = F}
# now gather all and plot with ggarrange

agg_png(filename = here::here("results","Fig2_micro.png"), res = 72, width = 1200, height = 1200)
ggarrange(rrrAUS, rrrAUT, rrrBEL, rrrBGR, rrrCAN, rrrCHE, rrrCHL, 
rrrCYP, rrrCZE, rrrDEU, rrrDNK, rrrESP, rrrFIN, rrrFRA, 
rrrGBR, rrrHUN, rrrIRL, rrrISR, rrrITA, rrrJPN, 
rrrLVA, rrrNLD, rrrNOR, rrrNZL, rrrPHL, rrrPOL, rrrPRT, 
rrrRUS, rrrSVK, rrrSVN, rrrSWE, rrrUSA, common.legend = TRUE) 
  
dev.off()


knitr::include_graphics(here::here("results","Fig2_micro.png"))
```

## Fig 4 and 5. Main. Correlation of Redist Variables and Context Vars

After observing the variation in the correlation between inclarge and incdiff by country. We suspect that attitudes toward the role of government are in play. We suspect these come from institutionalized norms. 

Theory: individuals have values or follow norms about how much the government should intervene in the socio-economic affairs of individuals, how much the government should protect against social risks and how much the government (society) should share or redistribute based on need. 

Hypothesis: In countries where the government does less redistribution or has a weaker welfare state, the correlation between perceptions of the income gap and the idea that the government is responsible for doing something about it will be smaller.


### Plot Liberal Business Ownership Corr (Liberal Values I)

```{r figx2a}
#agg_png(here::here("results","Fig2.png"), res = 144, height = 800, width = 1000)
Fig2.1 <- 
  df_macro %>%
  ggplot() +
 #geom_text(aes(y = corr, x = liberal_high, label = iso3c), color = "#20A387FF", size = 3) +
  geom_text(aes(y = corr, x = liberal, label = iso3c), color = "black", size = 3) +
  #geom_text(aes(y = corr, x = liberal_low, label = iso3c), color = "midnightblue", size = 3) +
  geom_text(aes(y = corr, x = ifelse(is.na(liberal), liberal_i, NA), label = iso3c), color = "grey", size = 2) +
  geom_smooth(aes(y = corr, x = liberal_i), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = liberal_high), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = liberal_low), method = 'lm', se = F, color = "midnightblue") +
  geom_segment(aes(x = 6.8, xend = 7.22, y = 0.30, yend = 0.30), linetype = "dashed", color = "#B8DE29FF", size = 1.1) +
  geom_segment(aes(x = 6.8, xend = 7.22, y = 0.27, yend = 0.27), color = "#20A387FF", size = 1.1) +
  geom_segment(aes(x = 6.8, xend = 7.22, y = 0.24, yend = 0.24), color = "midnightblue", size = 1.1) +
  annotate(geom = "text", x = 6.8, y = 0.33, label = "Fitted Lines", hjust = 0) +
  annotate(geom = "text", x = 7.3, y = 0.30, label = "Pooled", hjust = 0, size = 3) +
  annotate(geom = "text", x = 7.3, y = 0.27, label = "Higher GDP", hjust = 0, size = 3) +
  annotate(geom = "text", x = 7.3, y = 0.24, label = "Lower GDP", hjust = 0, size = 3) +
  labs(x = "LIBERAL VALUES I\nIntegrated Values Survey data\ngovernment v. private ownership of business",
       y = "CORRELATION within COUNTRY-YEAR\n'Government should redistribute' with\n'Differences in income are too large'"
       ) +
  theme_classic()
#dev.off()

#knitr::include_graphics(here::here("results","Fig2.png"))
```

### Plot Liberal Individual Welfare Corr (Liberal Values II)

Here Japan is an extreme outlier among rich countries, it has a very high level of support for the government being responsible rather than individuals. Contrast with Sweden which has extrememly high individual rather than government. This is likely a product of existing government arrangements. But Japan is so extreme that it makes sense to exclude it from the line of fit. 

```{r figx2}
#agg_png(here::here("results","Fig2.1_redist.png"), res = 144, height = 800, width = 1000)
Fig2.1_redist <- 
  df_macro %>%
  mutate(liberal_redist_high = ifelse(iso3c == "JPN", NA, liberal_redist_high)) %>%
  ggplot() +
  geom_text(aes(y = corr, x = liberal_redist, label = iso3c), color = "black", size = 3) +
  geom_text(aes(y = corr, x = ifelse(is.na(liberal_redist), liberal_redist_i, NA), label = iso3c), color = "grey", size = 2) +
  geom_smooth(aes(y = corr, x = liberal_redist_i), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = liberal_redist_high), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = liberal_redist_low), method = 'lm', se = F, color = "midnightblue") +
  # geom_segment(aes(x = 6.8, xend = 7.2, y = 0.3, yend = 0.3), linetype = "dashed", color = "#B8DE29FF", size = 1.1) +
  # geom_segment(aes(x = 6.8, xend = 7.2, y = 0.26, yend = 0.26), color = "#20A387FF", size = 1.1) +
  # geom_segment(aes(x = 6.8, xend = 7.2, y = 0.22, yend = 0.22), color = "midnightblue", size = 1.1) +
  # annotate(geom = "text", x = 6.8, y = 0.34, label = "Fitted Lines", hjust = 0) +
  # annotate(geom = "text", x = 7.3, y = 0.3, label = "Pooled", hjust = 0, size = 3) +
  # annotate(geom = "text", x = 7.3, y = 0.26, label = "Higher GDP", hjust = 0, size = 3) +
  # annotate(geom = "text", x = 7.3, y = 0.22, label = "Lower GDP", hjust = 0, size = 3) +
  labs(x = "LIBERAL VALUES II\nIntegrated Values Survey data\ngovernment v. individual provide for needs",
       y = ""
       ) +
  theme_classic()
#Fig2.1_redist
#dev.off()

#knitr::include_graphics(here::here("results","Fig2.1_redist.png"))
```

### Plot Corrupt / Corr

```{r figx3a}
#agg_png(here::here("results","Fig3.png"), res = 144, height = 800, width = 1000)
Fig2.2 <- df_macro %>%
  ggplot() +
  geom_text(aes(y = corr, x = cpiZ, label = iso3c), color = "black", size = 3) +
  geom_text(aes(y = corr, x = ifelse(is.na(cpiZ), cpiZ_i, NA), label = iso3c), color = "grey", size = 2) +
  geom_smooth(aes(y = corr, x = cpiZ), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = cpi_high), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = cpi_low), method = 'lm', se = F, color = "midnightblue") +
  geom_segment(aes(x = -2, xend = -1.7, y = 0.3, yend = 0.3), linetype = "dashed", color = "#B8DE29FF", size = 1.1) +
  geom_segment(aes(x = -2, xend = -1.7, y = 0.26, yend = 0.26), color = "#20A387FF", size = 1.1) +
  geom_segment(aes(x = -2, xend = -1.7, y = 0.22, yend = 0.22), color = "midnightblue", size = 1.1) +
  annotate(geom = "text", x = -2, y = 0.34, label = "Fitted Lines", hjust = 0) +
  annotate(geom = "text", x = -1.6, y = 0.3, label = "Pooled", hjust = 0, size = 3) +
  annotate(geom = "text", x = -1.6, y = 0.26, label = "Higher GDP", hjust = 0, size = 3) +
  annotate(geom = "text", x = -1.6, y = 0.22, label = "Lower GDP", hjust = 0, size = 3) +
  labs(x = "CORRUPTION PERCEPTIONS INDEX (CPI)\nTransparency International data\n ",
       y = "",
       #caption = " \nY data from International Social Survey Program\ngrey countries in left-panel interpolated with country-means"
       ) +
  theme_classic()

```

### Plot No Trust

```{r figx3}
#agg_png(here::here("results","Fig3.png"), res = 144, height = 800, width = 1000)
Fig2.3 <- df_macro %>%
  ggplot() +
  geom_text(aes(y = corr, x = noconf_govZ, label = iso3c), color = "black", size = 3) +
  geom_text(aes(y = corr, x = ifelse(is.na(noconf_govZ), noconf_govZ_i, NA), label = iso3c), color = "grey", size = 2) +
  geom_smooth(aes(y = corr, x = noconf_govZ), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = noconf_govZ_hi), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = noconf_govZ_low), method = 'lm', se = F, color = "midnightblue") +
  labs(x = "LACK OF TRUST IN GOV\nIntegrated Values Survey Data\n ",
       y = "",
       #caption = " \nY data from International Social Survey Program\ngrey countries in left-panel interpolated with country-means"
       ) +
  theme_classic()

```

### Individul-level corruption

Using the ISSP's own question on corruption in political careers.

```{r issp}
df_macro <- df_macro %>%
  mutate(corrupt_hi = ifelse(!is.na(gdp_hi), corrupt_top_r, NA),
         corrupt_lo = ifelse(!is.na(gdp_lo), corrupt_top_r, NA))

Fig2_micro <- df_macro %>%
  ggplot() +
    geom_text(aes(y = corr, x = corrupt_top_r, label = iso3c), color = "black", size = 3) +
  geom_smooth(aes(y = corr, x = corrupt_top_r), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = corrupt_hi), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = corrupt_lo), method = 'lm', se = F, color = "midnightblue") +
      labs(x = "Corruption in Political Careers\nISSP Data\n ",
       y = "",
       ) +
  theme_classic()

```


### Plot GDP

```{r figx4}
#agg_png(here::here("results","Fig3.png"), res = 144, height = 800, width = 1000)
Fig2.4 <- df_macro %>%
  ggplot() +
  geom_text(aes(y = corr, x = gdp_pc_10k, label = iso3c), color = "black", size = 3) +
  geom_smooth(aes(y = corr, x = gdp_pc_10k), method = 'lm', se = F, linetype = "dashed", color = "#B8DE29FF") +
  geom_smooth(aes(y = corr, x = gdp_hi), method = 'lm', se = F, color = "#20A387FF") +
  geom_smooth(aes(y = corr, x = gdp_lo), method = 'lm', se = F, color = "midnightblue") +
  #geom_segment(aes(x = -2, xend = -1.7, y = 0.3, yend = 0.3), linetype = "dashed", color = "#B8DE29FF", size = 1.1) +
  #geom_segment(aes(x = -2, xend = -1.7, y = 0.26, yend = 0.26), color = "#20A387FF", size = 1.1) +
  #geom_segment(aes(x = -2, xend = -1.7, y = 0.22, yend = 0.22), color = "midnightblue", size = 1.1) +
  # annotate(geom = "text", x = -2, y = 0.34, label = "Fitted Lines", hjust = 0) +
  # annotate(geom = "text", x = -1.6, y = 0.3, label = "Pooled", hjust = 0, size = 3) +
  # annotate(geom = "text", x = -1.6, y = 0.26, label = "Higher GDP", hjust = 0, size = 3) +
  # annotate(geom = "text", x = -1.6, y = 0.22, label = "Lower GDP", hjust = 0, size = 3) +
  labs(x = "GDP PER CAPITA (10k)\nVarieties of Democracy data\n ",
       y = "",
       #caption = " \nY data from International Social Survey Program\ngrey countries reflect interpolated data points"
       ) +
  theme_classic()

```
### Plot Arranged Figure

```{r fig2}
agg_png(filename = here::here("results","Fig4.png"), res = 144, width = 1250, height = 1250)
ggarrange(Fig2.2, Fig2_micro, Fig2.3, Fig2.4, 
          widths = c(1,1,1,1), 
          nrow = 2, ncol=2)
dev.off()

knitr::include_graphics(here::here("results","Fig4.png"))
```

```{r fig3}
agg_png(filename = here::here("results","Fig5.png"), res = 144, width = 2000, height = 800) +
  theme(text = element_text())
ggarrange(Fig2.2, Fig2_micro, widths = c(1.1, 1, 1), nrow=1)
dev.off()

knitr::include_graphics(here::here("results","Fig5.png"))
```



## References

```{r cite}
report::cite_packages()
```


## Colophon

```{r colophon}
sessionInfo()
```
