==> 05_Analysis.Rmd <==
---
title: "05 Analysis"
output:
  html_document:
    df_print: paged
---

##### ISSP, Social Inequality, Cumulation (1987, 1992, 1999, 2009 & 2019)








```{r setup, include=FALSE}
library(pacman)
pacman::p_load("tidyverse", "ggpubr", 
               "ragg", "kableExtra", 
               "lme4", "ggrepel", 
               "sjPlot",
               "arm",
               "margins",
               "ggeffects",
               "modelsummary")

knitr::opts_chunk$set(message = F, warning = F)

```


```{r load dfs}
#df <- readRDS(here::here("data","df.RDS"))
df_na <- readRDS(here::here("data", "df_na.RDS"))
df_macro <- readRDS(here::here("data", "df_macro.RDS"))

```

### Macro-Correlations

```{r macro_corrs}
macro_corr <- df_macro %>% 
  dplyr::select(gov_redist, incdiff_large, corrupt_top_r, corr, 
         libZ, libZ_i, lib_redistZ, lib_redistZ_i, 
         noconf_govZ, noconf_govZ_i, cpiZ, cpiZ_i, 
         gdp_pc_10k, socx, pct_fb_i, gini_disp) %>%
  cor(., use = "pairwise.complete.obs") %>%
  round(2)

macro_corr[upper.tri(macro_corr)] <- ""

write.csv(macro_corr, here::here("results","Tbl2.csv"), row.names = F)
```


### All Plausible Models

#### Moderation models, for specification curve

Selection point. If we've run the models and the margins command before then this will not be necessary. 

NOTE: that if we make changes to the models we have to turn off the selector and re-run, or delete the df_ame file, or the if selection won't work!

```{r ameload}
if (file.exists(here::here("data", "df_ame.RDS"))) { 
  # select if this margins datafile is already there

   df_ame <- readRDS(here::here("data", "df_ame.RDS"))

    } else 
 { source(here::here("data", "Models_nomain.R"))
   model_list <- mget(ls(pattern = "M_"))
   #model_list_lm <- mget(ls(pattern = "L_"))
#########
   # Compile all M_ models margins (multilevel)   
   df_ame <- as.data.frame(matrix(nrow = length(model_list), ncol = 3))

colnames(df_ame) <- c("Model", "mod_b", "mod_se")

model_names <- names(model_list)

for (m in 1:length(model_list)) {
  
  #mn <- as.name(model_names[m])
  
  df_ame[m,1] <- paste(as.character(model_list[[m]]@call), sep = "", collapse = "")
  
  temp <- as.data.frame(summary(get(model_names[m]))[["coefficients"]])
  
  temp <- temp[grepl('cpiZ|noconf_govZ|libZ|corrupt_top|lib_redistZ|incdiff_large|gdp_pc_10k_C', rownames(temp)), 1:2]
  
  df_ame[m,"mod_b"] <- temp[grepl(":", rownames(temp)),1]
  df_ame[m, "mod_se"] <- temp[grepl(":", rownames(temp)),2]
  
}


}
```

#### Duplicates

There is a bug in the routine that leads to a handful of duplicated models for GDP as the moderator. Remove them here

```{r dups}
df_ame <- df_ame[!duplicated(df_ame[c("mod_b", "mod_se")]),]

```


#### Get model specifications

```{r ameclean}
df_ame <- df_ame %>%
  mutate(mod_var = ifelse(grepl("noconf_govZ", Model, fixed = T), "B_No_confidence", 
                   ifelse(grepl("cpiZ", Model), "A2_Corruption_percept",
                   ifelse(grepl("corrupt_top", Model), "A1_Corrupt_micro",
         ifelse(grepl("gdp_pc_10k_C:", Model), "C_GDP", NA)))),
         mod_imputed = ifelse(grepl("cpiZ_i|noconf_govZ_i", df_ame$Model), "Yes", "No"),
         ri_cyear = ifelse(grepl("iso3c_wave", df_ame$Model), "Yes", "No"),
         ri_country = ifelse(grepl("iso3c)", df_ame$Model), "Yes", "No"),
         estimation = ifelse(grepl("lmer", df_ame$Model), "GLM", "OLS"),
         incdiff_grand_mean = ifelse(grepl("incdiff_large_C", df_ame$Model), "Yes", "No"),
         incdiff_within_cyear = ifelse(grepl("incdiff_large_w", df_ame$Model), "Yes", "No"),
         control_gdp = ifelse(grepl("gdp", df_ame$Model), "Yes", "No"),
         control_gini = ifelse(grepl("gini", df_ame$Model), "Yes", "No"),
         control_socx = ifelse(grepl("socx", df_ame$Model), "Yes", "No"),
         control_fb = ifelse(grepl("pct_fb", df_ame$Model), "Yes", "No"),
         i_level_control = ifelse(grepl("female", df_ame$Model), "Yes", "No"),
         gdp_m_interaction = ifelse(grepl(":gdp_pc_10k_C", df_ame$Model), "Yes", "No"),
         gdp_m_interaction = ifelse(mod_var == "C_GDP", "Yes", gdp_m_interaction),
         redundant = ifelse(grepl("gdp_pc_10k_Cdata", df_ame$Model), "Yes", "No")
         )

df_ame <- df_ame %>%
  arrange(-mod_b) %>%
  mutate(count = row_number())

saveRDS(df_ame, here::here("data", "df_ame.RDS"))

```


## Percentage of Agree

Here we compare our heuristics with the percentage of those who do not agree the government should redistribute out of those who agree inequality is too high (the 'inconsistent'
attitude individuals). We also compare them based on the correlation because the inconsistent too much inequality perceivers do not pattern by two of three heuristics.

  

### Setup Data

```{r pctagree}
df_tbl <- df_macro %>%
  dplyr::select(iso3c, wave, corr, gap_no_gov, gap_no_gov_opposite, cpiZ_i, noconf_govZ_i, gdp_pc_10k) %>%
  group_by(iso3c,wave) %>%
  summarise_all(mean, na.rm = T) %>%
  ungroup()

df_tbl <- df_tbl %>%
  mutate(gap_no_gov = round(gap_no_gov,1),
         wavea = ifelse(wave > 1999, wave - 2000, wave - 1900),
         wavea = ifelse(wavea == 9, "09", wavea),
         clabs = paste0(iso3c,"'",wavea))
```


### Figure 5


#### CPI

##### Inconsistent

```{r cpi_inc}
p1 <- df_tbl %>%
  ggplot(aes(cpiZ_i, gap_no_gov)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#73D055FF") +
  labs(x = "",
       y = "% Inconsistent Attitudes\nDo not agree with gov redistribution") +
  theme_classic()


# df_tbl %>%
#   ggplot(aes(cpiZ_i, gap_no_gov_opposite)) +
#   geom_text_repel(aes(label = clabs)) +
#   geom_smooth(method = lm)
  
```

##### Correlation

```{r cpi_corr}
p4 <- df_tbl %>%
  ggplot(aes(cpiZ_i, corr)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#73D055FF") +
  labs(x = "Corruption Perceptions Index",
       y = "Correlation of Percieved Inequality\nand Support for Redistribution") +
  theme_classic()
```



#### No Trust

##### Inconsistent

```{r trust_inc}
p2 <- df_tbl %>%
  ggplot(aes(noconf_govZ_i, gap_no_gov)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#287D8EFF") +
  labs(x = "",
       y = "") +
  theme_classic()

# df_tbl %>%
#   ggplot(aes(noconf_govZ_i, gap_no_gov_opposite)) +
#   geom_text_repel(aes(label = clabs)) +
#   geom_smooth(method = lm)
```

##### Correlation

```{r trust_corr}
p5 <- df_tbl %>%
  ggplot(aes(noconf_govZ_i, corr)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#287D8EFF") +
  labs(x = "Lack of Trust in Government",
       y = "") +
  theme_classic()
```

#### GDP

##### Inconsistent
```{r gdp_inc}
p3 <- df_tbl %>%
  ggplot(aes(gdp_pc_10k, gap_no_gov)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#453781FF") +
  labs(x = "",
       y = "") +
  theme_classic()

# df_tbl %>%
#   ggplot(aes(gdp_pc_10k, gap_no_gov_opposite)) +
#   geom_text_repel(aes(label = clabs)) +
#   geom_smooth(method = lm)
```
##### Correlation

```{r gdp_corr}
p6 <- df_tbl %>%
  ggplot(aes(gdp_pc_10k, corr)) +
  geom_text_repel(aes(label = clabs), size = 3) +
  geom_smooth(method = lm, color = "#453781FF") +
  labs(x = "Worse Economic Performance (GDPpc)",
       y = "") +
  theme_classic()
```

#### Arrange

```{r 6way}

ggarrange(p1,p2,p3,p4,p5,p6, nrow = 2, ncol = 3, 
          widths = c(1.05,1,1,1.05,1,1), 
          heights = c(1,1,1,1.1,1.1,1.1)) %>%
  ggexport(filename = here::here("results", "Fig5.png"), res = 144, width = 1750, height = 1100)

knitr::include_graphics(here::here("results", "Fig5.png"))
```

## References

```{r cite}
report::cite_packages()
```


## Colophon

```{r colophon}
sessionInfo()
```
