==> 07_Weisstanner_Replication.Rmd <==
---
title: "07 Weisstanner Replication"
output:
  word_document: default
  html_document:
    df_print: paged
---

An earlier version of the paper included a replication of Weisstanner, we leave it in here for completeness. 

### Weisstanner Replication

Obtained original Stata code from author.
Altered do files to extract the variable 'incdiff' measuring 'attitudes toward inequality' from each Social Inequality ISSP module

cr_ISSP_1987.do

cr_ISSP_1992.do

cr_ISSP_1999.do

cr_ISSP_2009.do

Added ISSP source .dta files, changed code to use these files instead of Weisstanner's naming convention, e.g., "ISSP 1992 Inequality - ZA2310_SPSS.dta" in the 'use' command would be changed to simply "ZA2310.dta"

We removed or commented out code to extract ESS data or non-Social Inequality ISSP module data. We downloaded the necessary ISSP datasets from GESIS.

We renamed file paths.

We changed the final dataset name from AbsRel_FINAL.dta to AbsRel_FINAL_REP.dta.

We created a variable with all missing values for the variables redist1 & redist4 to allow the code to run.

This affected the following code:

AbsRel_0.3_Finalise_LIS_estimates_CODE.do
AbsRel_1_ESS and ISSP Datasets.do
AbsRel_2_Label variables.do
AbsRel_3_Models.do

Then we ran the above two code chunks in that order. We can confirm that these models produce the original results, even on the smaller sample. The confidence intervals are not surprisingly much larger.



```{r setup}
pacman::p_load("tidyverse", 
               "countrycode",
               "readstata13",
               "lme4",
               "sjPlot",
               "ragg")

options(scipen = 999, digits = 3)
knitr::opts_chunk$set(message = F, warning = F)
```

## Data

```{r loaddata}
df_macro <- readRDS(here::here("data", "df_macro.RDS"))
df_na <- readRDS(here::here("data", "df_na.RDS"))
df_rep <- readstata13::read.dta13(here::here("weisstanner_replication","REPLICATION DATA","AbsRel_FINAL_REP.dta"))

# select out heuristic vars
df_macro <- df_macro %>%
  dplyr::select(iso3c, wave, libZ_i, lib_redistZ_i, noconf_govZ_i, cpiZ_i)
#merge
# NA and age selection

df_rep <- df_rep %>%
  subset(!is.na(redist3) & age > 17 & age < 66) %>%
  mutate(iso3c = countrycode(country, "country.name", "iso3c"),
         redist_ineq = ifelse(incdiff == "Strongly agree", 5, ifelse(incdiff == "Agree", 4, ifelse(incdiff == "Neither nor", 3, ifelse(incdiff == "Disagree", 2, ifelse(incdiff == "Strongly disagree", 1, NA)))))) %>%
  left_join(df_macro, by = c("iso3c","wave"))

```


## Original Model

Results are identical as those when running the xtreg Stata command using the modified original code

```{r omodel}

m02 <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

m03 <- lmer(redist_dum ~ relgr5_wa + absgr5_wa*hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

#summary(m02)
```

## Original Model Stripped

We test removing the country level variables given that we now have only 36 country-wave units. Results for absolute and relative are nearly identical, therefore we proceed with this model for parsimony. We remove GDP also (see note in next chunk)

```{r omodel_strip}
m02s <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

#summary(m02s)
```

Replicated Results
```{r mouts}
tab_model(m02, m03, p.style = "stars", show.ci = F)
```

## Heuristics Added

## Models

### First Association

We first observe that government heuristics associate with support for government redistribution.

We also remove GDP because these are all rich countries, and we've seen in both our models and in his models, that GDP is not a strong predictor, but it introduces endogeneity due to its correlation with our heuristic variables. 

Something striking is that people in more corrupt societies are more likely to support government redistribution.

#### Models

```{r hmodel}
m04_1 <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + libZ_i + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i))) 

m04_2 <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + lib_redistZ_i + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

m04_3 <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + noconf_govZ_i + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

m04_4 <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + cpiZ_i + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

```

#### Results

```{r hmodeltab}
tab_model(m04_1, m04_2, m04_3, m04_4, p.style = "stars", show.ci = F)
```

### Add Att to Inequality

This of course removes the effects we aim to recover, but demonstrates that even when absolute and relative are in the model, we can still observe the impact of government heuristics (negative in all cases for these rich countries)

#### Models

```{r hmodeli}
m04_1i <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + redist_ineq*libZ_i + (1 | cwave), data = df_rep) 

m04_2i <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + redist_ineq*lib_redistZ_i + (1 | cwave), data = df_rep)

m04_3i <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + redist_ineq*noconf_govZ_i + (1 | cwave), data = df_rep)

m04_4i <- lmer(redist_dum ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + redist_ineq*cpiZ_i + (1 | cwave), data = df_rep)

```

#### Results

```{r hmodeltabi}
tab_model(m04_1i, m04_2i, m04_3i, m04_4i, p.style = "stars", show.ci = F)
```

### Counterfactual Gov Support

We now calculate the predicted association based on heuristic and att to inequality.

#### Prediction Models

```{r hmodelp}

m04_1p <- lmer(redist_dum ~ redist_ineq*libZ_i + (1 | cwave), data = df_rep)

m04_2p <- lmer(redist_dum ~ redist_ineq*lib_redistZ_i + (1 | cwave), data = df_rep)

m04_3p <- lmer(redist_dum ~ redist_ineq*noconf_govZ_i + (1 | cwave), data = df_rep)

m04_4p <- lmer(redist_dum ~ redist_ineq*cpiZ_i + (1 | cwave), data = df_rep)
# create dataframe where heuristics are lowest

df_rep_x <- df_rep %>%
  mutate(libZ_i = ifelse(!is.na(libZ_i), -2, NA),
         lib_redistZ_i = ifelse(!is.na(lib_redistZ_i), -2, NA),
         noconf_govZ_i = ifelse(!is.na(noconf_govZ_i), -2, NA),
         cpiZ_i = ifelse(!is.na(cpiZ_i), -1.5, NA))

# get predicted values for normal and low lib scores
df_rep$redist_dum_a1_liblo <- predict(m04_1p, newdata = df_rep_x, na.action = "na.exclude")
df_rep$redist_dum_a1 <- predict(m04_1p, newdata = df_rep, na.action = "na.exclude")

df_rep$redist_dum_a2_liblo <- predict(m04_2p, newdata = df_rep_x, na.action = "na.exclude")
df_rep$redist_dum_a2 <- predict(m04_2p, newdata = df_rep, na.action = "na.exclude")

df_rep$redist_dum_b_liblo <- predict(m04_3p, newdata = df_rep_x, na.action = "na.exclude")
df_rep$redist_dum_b <- predict(m04_3p, newdata = df_rep, na.action = "na.exclude")

df_rep$redist_dum_c_liblo <- predict(m04_4p, newdata = df_rep_x, na.action = "na.exclude")
df_rep$redist_dum_c <- predict(m04_4p, newdata = df_rep, na.action = "na.exclude")

#trim into categories
df_rep <- df_rep %>%
  #lib values I normal
  mutate(
         #lib values I normal
         redist_dum_a1 = ifelse(redist_dum_a1 < 0, 0, ifelse(redist_dum_a1 > 1, 1, redist_dum_a1)),
         redist_dum_lib = round(redist_dum_a1, 0),
         #lib values I low
         redist_dum_a1_liblo = ifelse(redist_dum_a1_liblo < 0, 0, ifelse(redist_dum_a1_liblo > 1, 1, redist_dum_a1_liblo)),
         redist_dum_lib_low = round(redist_dum_a1_liblo, 0),
         #lib values II normal
         redist_dum_a2 = ifelse(redist_dum_a2 < 0, 0, ifelse(redist_dum_a2 > 1, 1, redist_dum_a2)),
         redist_dum_lib_redist = round(redist_dum_a2, 0),
         #lib values II low
         redist_dum_a2_liblo = ifelse(redist_dum_a2_liblo < 0, 0, ifelse(redist_dum_a2_liblo > 1, 1, redist_dum_a2_liblo)),
         redist_dum_lib_redist_low = round(redist_dum_a2_liblo, 0),
         #lack trust normal
         redist_dum_b = ifelse(redist_dum_b < 0, 0, ifelse(redist_dum_b > 1, 1, redist_dum_b)),
         redist_dum_notrust = round(redist_dum_b, 0),
         #lack trust low
         redist_dum_b_liblo = ifelse(redist_dum_b_liblo < 0, 0, ifelse(redist_dum_b_liblo > 1, 1, redist_dum_b_liblo)),
         redist_dum_notrust_low = round(redist_dum_b_liblo, 0),
         #cpi normal
         redist_dum_c = ifelse(redist_dum_c < 0, 0, ifelse(redist_dum_c > 1, 1, redist_dum_c)),
         redist_dum_cpi = round(redist_dum_c, 0),
         #cpi low
         redist_dum_c_liblo = ifelse(redist_dum_c_liblo < 0, 0, ifelse(redist_dum_c_liblo > 1, 1, redist_dum_c_liblo)),
         redist_dum_cpi_low = round(redist_dum_c_liblo, 0))

```

#### Replication Models

these models should confirm that we can accurately proxy the original measure and come to similar results

```{r expansion}
m02_rep_a1 <- lmer(redist_dum_lib ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = df_rep)

m02_rep_a2 <- lmer(redist_dum_lib_redist ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = df_rep)

m02_rep_b <- lmer(redist_dum_notrust ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = df_rep)

m02_rep_c <- lmer(redist_dum_cpi ~ absgr5_wa*relgr5_wa + hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = df_rep)
```

#### Results

```{r hmodeltabp}
tab_model(m02_rep_a1, m02_rep_a2, m02_rep_b, m02_rep_c, p.style = "stars", show.ci = F)
```


### M03 with new DV
```{r replication_new}
# for decile figure
m03lib <- lmer(redist_dum_lib ~ relgr5_wa + absgr5_wa*hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

m03liblo <- lmer(redist_dum_lib_low ~ relgr5_wa + absgr5_wa*hinc_decile + factor(educ_tert) + female + age + factor(unemployed) + factor(wave) + unemp + socexp + gini_mkt + gdppc + (1 | cwave), data = subset(df_rep, !is.na(df_rep$libZ_i)))

tab_model(m03, m03lib, m03liblo, p.style = "stars", show.ci = F)
```
## Compare Margins

We use Stata to compare our new marginal graphs with those of the original study.

See file "AbsRel_4_Models_Adjusted.do"

```{r stata_export}
#make smaller
# fix a change in naming convention
df_rep$redist_dum_libZ_i <- df_rep$redist_dum_lib_low
df_rep$redist_dum_libZ_redist_i <- df_rep$redist_dum_lib_redist_low
# note that notrust and cpi are now the original rather than the low values, we are not using them in the paper, but they would need to be changed by name here to use them

df_rep <- df_rep %>%
  select(c(redist_dum_libZ_i, redist_dum, redist_dum_libZ_redist_i, redist_dum_notrust, redist_dum_cpi, absgr5_wa, relgr5_wa, hinc_decile, educ_tert, female, age, unemployed, wave, unemp, socexp, gini_mkt, gdppc, incdiff, iso3c, cwave))

save.dta13(df_rep, here::here("weisstanner_replication", "REPLICATION DATA", "AbsRel_ADJUSTED.dta"))

rm(df_rep_x)
```

Case study USA and FIN

Liberal values are very strong in USA, we can build a counterfactual to see how absolute and relative inequality affects income deciles.
In the margins function we calculate average marginal effects using the abs and rel income by decile but applying it to the entire dataset.

### Call margins function

```{r marginfunc}
source(here::here("weisstanner_replication", "margin_function.R"))
```

                         
#### Fig 6A.1 Original Margins

```{r plot6A}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6a1.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() + 
  #USA
  geom_point(aes(x = factor(hinc_decile), y = redist_o_obs_obs_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_obs_obs_m-(redist_o_obs_obs_sd), ymax = redist_o_obs_obs_m+(redist_o_obs_obs_sd), color = country), width = 0.2) +
  theme_classic() +
    annotate("text", x = 3, y = .25, label = "U.S.", color = "#3CBB75FF") +
  annotate("text", x = 8, y = .83, label = "Finland", color = "#39568CFF") +
  labs(x = "Income Decile", y = "Average Marginal\nSupport for Redistribution", color = "Liberal Values I") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
  scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  theme(
    legend.position = "none"
  )
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6a1.png"))
```

#### Fig 6A.2 Original Stagnated Margins

```{r plot6A2}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6a2.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() + 
  #USA
  geom_point(aes(x = factor(hinc_decile), y = redist_o_stag_obs_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_stag_obs_m-(redist_o_stag_obs_sd), ymax = redist_o_stag_obs_m+(redist_o_stag_obs_sd), color = country), width = 0.2) +
  theme_classic() +
  labs(x = "Income Decile", y = "Average Marginal\nSupport for Redistribution", color = "Liberal Values I") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
  scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  theme(
    legend.position = "none"
  )
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6a2.png"))
```
#### Fig 6B.1 Adjusted Predictions, Original

```{r plot_usa_lo}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6b1.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() +
  #USA
  geom_point(aes(x = factor(hinc_decile), y = redist_o_obs_lib_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_obs_lib_m-(redist_o_obs_lib_sd), ymax = redist_o_obs_lib_m+(redist_o_obs_lib_sd), color = country), width = 0.2) +
  theme_classic() +
  labs(x = "Income Decile", y = "", color = "Liberal\nValues I") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
  scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  guides(color = guide_legend(nrow = 2)) +
  theme(legend.position = "none")
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6b1.png"))
```
#### Fig 6B.2 Adjusted Predictions, Stagnated

```{r plot_fin}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6b2.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() +
  #FIN 
  geom_point(aes(x = factor(hinc_decile), y = redist_o_stag_lib_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_stag_lib_m-(redist_o_stag_lib_sd), ymax = redist_o_stag_lib_m+(redist_o_stag_lib_sd), color = country), width = 0.2) +
  theme_classic() +
  ylim(0.15,1) +
  labs(x = "Income Decile", y = "", color = "Liberal Values I") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
    scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  theme(
    legend.position = "none"
  )
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig6b2.png"))

```
### Fig 7 Adjusted Predictions, Abs Obs, Lib Low

```{r plot_fin_lo}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig7a.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() +
  #USA 
  geom_point(aes(x = factor(hinc_decile), y = redist_o_obs_liblow_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_obs_liblow_m-(redist_o_obs_liblow_sd), ymax = redist_o_obs_liblow_m+(redist_o_obs_liblow_sd), color = country), width = 0.2) +
  theme_classic() +
  labs(x = "Income Decile", y = "Average Marginal\nSupport for Redistribution", color = "Liberal\nValues I") +
  annotate("text", x = 5, y = .5, label = "U.S.", color = "#3CBB75FF") +
  annotate("text", x = 8, y = .98, label = "Finland", color = "#39568CFF") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
  scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  guides(color = guide_legend(nrow = 2)) +
  theme(legend.position = "none",
        legend.box = "vertical") 
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig7a.png"))

```

### Fig 7 Adjusted Predictions, Abs Low, Lib Low

```{r plot_fin_lo}
agg_png(here::here("weisstanner_replication","REPLICATION DATA","results","Fig7b.png"), width = 550, height = 350, res = 144)
df_rep_marg%>%
  ggplot() +
  #USA 
  geom_point(aes(x = factor(hinc_decile), y = redist_o_stag_liblow_m, color = country)) +
  geom_errorbar(aes(x = factor(hinc_decile), ymin = redist_o_stag_liblow_m-(redist_o_stag_liblow_sd), ymax = redist_o_stag_liblow_m+(redist_o_stag_liblow_sd), color = country), width = 0.2) +
  theme_classic() +
  labs(x = "Income Decile", y = "", color = "Liberal\nValues I") +
  scale_color_manual(values = c("#39568CFF",
                                "#3CBB75FF")) +
  scale_y_continuous(labels = scales::percent,
                     limits = 0:100) +
  guides(color = guide_legend(nrow = 2)) +
  theme(legend.position = "none",
        legend.box = "vertical") 
dev.off()

knitr::include_graphics(here::here("weisstanner_replication","REPLICATION DATA","results","Fig7b.png"))

```
